refactor(NonVerfiedMailMessage): Extract NonVerfiedMailMessage to own comp
